+++ src/actors/triangle.rs
@@ -395,6 +395,12 @@ impl Enemy for Triangle{
         }
         return None
     }
+
+    fn get_all_draw_calls(&self) -> Vec<DrawCall>{
+        let selfcal = self.get_draw_call();
+
+        return vec![selfcal]
+    }
 }
 
 #[async_trait]