+++ src/event_system/interface.rs
@@ -61,6 +61,8 @@ pub trait Enemy: GameEntity{
 
     fn force_state(&mut self, state: StateType);
     fn get_state(&self) -> Option<StateType>;
+
+    fn get_all_draw_calls(&self) -> Vec<DrawCall>;  //REVIEW: Currently only serves debugging purposes.
 }
 
 #[async_trait]
@@ -72,5 +74,6 @@ pub trait Projectile: GameEntity{
 
     fn force_state(&mut self, state: StateType);
     fn get_state(&self) -> Option<StateType>;
+    
     fn get_all_draw_calls(&self) -> Vec<DrawCall>;  //REVIEW: Currently only serves debugging purposes.
 }